build script: only upload if running on openframeworks/PG/master

+++ scripts/osx/buildPG.sh
@@ -11,9 +11,12 @@ if [ $ret -ne 0 ]; then
       echo "Failed building Project Generator"
       exit 1
 fi
-mv projectGeneratorSimple/bin/data/settings/projectGeneratorSettings_production.xml projectGeneratorSimple/bin/data/settings/projectGeneratorSettings.xml
-cp scripts/ssh_config ~/.ssh/config
-chmod 600 scripts/id_rsa
-ssh -i scripts/id_rsa tests@192.237.185.151 "rm -r projectGeneratorSimple_osx"
-scp -i scripts/id_rsa -r projectGeneratorSimple/bin tests@192.237.185.151:projectGeneratorSimple_osx
+if [ "${TRAVIS_REPO_SLUG}/${TRAVIS_BRANCH}" = "openframeworks/projectGenerator/travis"]; then
+    mv projectGeneratorSimple/bin/data/settings/projectGeneratorSettings_production.xml projectGeneratorSimple/bin/data/settings/projectGeneratorSettings.xml
+    cp scripts/ssh_config ~/.ssh/config
+    chmod 600 scripts/id_rsa
+    ssh -i scripts/id_rsa tests@192.237.185.151 "rm -r projectGeneratorSimple_osx"
+    scp -i scripts/id_rsa -r projectGeneratorSimple/bin tests@192.237.185.151:projectGeneratorSimple_osx
+fi
 rm scripts/id_rsa
+
